Refactor: step1 복잡한 hook 분리
